requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
groq==0.4.2
mysql-connector-python==8.2.0
pydantic==2.5.0
pandas==2.2.0
numpy==2.2.0
scikit-learn==1.6.0
xgboost==3.2.0
python-multipart==0.0.6
